# Refactor Arkivo::Actor spec