--- a/spec/lib/hyrax/arkivo/actor_spec.rb
+++ b/spec/lib/hyrax/arkivo/actor_spec.rb
@@ -20,42 +20,26 @@ describe Hyrax::Arkivo::Actor do
     it { is_expected.to respond_to(:content_type=) }
   end
 
+  let(:work_actor) { instance_double(Hyrax::Actors::ActorStack) }
+  let(:file_actor) { double }
+  before do
+    allow(Hyrax::CurationConcern).to receive(:actor).and_return(work_actor)
+    allow(Hyrax::Actors::FileSetActor).to receive(:new).and_return(file_actor)
+  end
+
   describe '#create_work_from_item' do
-    it { is_expected.to respond_to(:create_work_from_item) }
+    it 'creates a work and a file and returns a GenericWork' do
+      expect(work_actor).to receive(:create).with(
+        hash_including(arkivo_checksum: item['file']['md5'],
+                       "title" => [item['metadata']['title']])
+      ).and_return(true)
+      expect(file_actor).to receive(:create_metadata)
+      expect(file_actor).to receive(:create_content) do |tmpfile|
+        expect(tmpfile).to be_instance_of Tempfile
+        expect(tmpfile.read).to eq "arkivo\n"
+      end
+      expect(file_actor).to receive(:attach_file_to_work)
 
-    it 'instantiates an actor' do
-      expect(Hyrax::Actors::GenericWorkActor).to receive(:new).once.and_call_original
-      subject.create_work_from_item
-    end
-
-    it 'creates initial metadata' do
-      expect_any_instance_of(Hyrax::Actors::FileSetActor).to receive(:create_metadata).once
-      subject.create_work_from_item
-    end
-
-    it 'stores a checksum' do
-      work = subject.create_work_from_item
-      expect(work.arkivo_checksum).to eq item['file']['md5']
-    end
-
-    it 'calls create_content' do
-      expect_any_instance_of(Hyrax::Actors::FileSetActor).to receive(:create_content).once
-      subject.create_work_from_item
-    end
-
-    it 'extracts a file from the item' do
-      work = subject.create_work_from_item
-      expect(work.reload.file_sets.first.original_file.content).to eq "arkivo\n"
-    end
-
-    it 'batch applies metadata' do
-      work = subject.create_work_from_item
-      # TODO: Figure out why this is needed if the Resque job is running synchronously
-      reloaded = work.reload
-      expect(reloaded.title).to eq Array.wrap(item['metadata']['title'])
-    end
-
-    it 'returns a GF instance' do
       expect(subject.create_work_from_item).to be_instance_of(GenericWork)
     end
   end
@@ -71,64 +55,24 @@ describe Hyrax::Arkivo::Actor do
         f.arkivo_checksum = checksum
       end
     end
-    let(:file_set) { create(:file_set, user: user) }
 
-    before do
-      work.ordered_members << file_set
-      work.save!
-    end
-
-    it { is_expected.to respond_to(:update_work_from_item) }
-
-    it 'instantiates an actor' do
-      expect(Hyrax::Actors::FileSetActor).to receive(:new).once.and_call_original
-      subject.update_work_from_item(work)
-    end
-
-    describe '#reset_metadata' do
-      it 'changes the title' do
-        # For some reason, "expect to change from to" wasn't working here
-        expect(work.title).to eq title
-        subject.update_work_from_item(work)
-        expect(work.title).to eq Array.wrap(item['metadata']['title'])
+    it 'changes the title and clears other metadata' do
+      expect(work_actor).to receive(:update).with(hash_including("title" => [item['metadata']['title']],
+                                                                 "description" => [],
+                                                                 arkivo_checksum: item['file']['md5']))
+      expect(file_actor).to receive(:update_content) do |tmpfile|
+        expect(tmpfile).to be_instance_of Tempfile
+        expect(tmpfile.read).to eq "# HEADER\n\nThis is a paragraph!\n"
       end
-
-      it 'wipes out the description' do
-        # For some reason, "expect to change from to" wasn't working here
-        expect(work.description).to eq description
-        subject.update_work_from_item(work.reload)
-        expect(work.description).to eq []
-      end
-    end
-
-    it 'changes the arkivo checksum' do
-      expect do
-        subject.update_work_from_item(work)
-      end.to change { work.arkivo_checksum }.from(checksum).to(item['file']['md5'])
-    end
-
-    it 'calls update_content' do
-      expect_any_instance_of(Hyrax::Actors::FileSetActor).to receive(:update_content).once
-      subject.update_work_from_item(work)
-    end
-
-    # TODO: this is testing FileSetActor, so it is not needed here.
-    it 'extracts a file from the item' do
-      subject.update_work_from_item(work)
-      expect(file_set.reload.original_file.content).to eq "# HEADER\n\nThis is a paragraph!\n"
-    end
-
-    it 'returns a GF instance' do
       expect(subject.update_work_from_item(work)).to be_instance_of(GenericWork)
     end
   end
 
   describe '#destroy_work' do
-    let(:work) { create(:generic_work, user: user) }
+    let(:work) { mock_model(GenericWork) }
     it 'deletes the file' do
-      expect do
-        subject.destroy_work(work)
-      end.to change { work.destroyed? }.from(nil).to(true)
+      expect(work).to receive(:destroy)
+      subject.destroy_work(work)
     end
   end
 end