Fixes #16 

@projecthydra/hyrax-code-reviewers

Previously it was testing the whole of the stack, now it just tests the
class under test. The new version is 33x faster:

Before:
```
Finished in 22.72 seconds (files took 5.01 seconds to load)
```

After:
```
Finished in 0.6944 seconds (files took 5.02 seconds to load)
```