@mjgiarlo 2016-12-12
@jeremyf I believe @jcoyne has made the change you requested. Mind re-reviewing?